11-21-16. FindAltPath and run changes.
wrote the code to actually have the robots and obstacles move. Wrote the
findAlernatePath() method on robot. added methods to check if the Robot
ether made it to the destination or if it hit an obstacle(gameOver()).
Robot no longer updates that its position is occupied on the grid for
the gameOver method which checks to see if the vertice at the Robot's
position is occupied, and if it is, then an obstacle must be there and
the game is over. 

Made some test objects and path to see if the move methods were working
successfully.  

Added place holder findPath(), gameLost, gameWon, and
printGameStateMethods().