Fixed Loop porblem hopefully.
it now exits if the destination is occupied for more than 4 turns in a
row(at least it is suppose to. Seems to only work if there is a
stationary obstacle on the destination). Also as a safegaurd added the
game to quit after ten times as many turns as the size of the grid
because it definitely should not take that many turns unless it is in a
loop.